Cleaning up ReactUWP's IDLs and WinRT code
