+++ RNWCPP/ReactUWP/ABI/ReactControl_rt.cpp
@@ -5,6 +5,7 @@
 
 #include "comTemplateLibrary.h"
 #include "HStringHelper.h"
+#include "Instance_rt.h"
 #include "ReactControl_rt.h"
 #include <Windows.Foundation.h>
 #include <windows.ui.xaml.controls.h>
@@ -15,112 +16,6 @@ namespace react {
 namespace uwp {
 
 ActivatableClassWithFactory(ReactControl, ReactControlFactory);
-ActivatableClassWithFactory(Instance, InstanceStatics);
-
-struct InstanceReactInstanceCreator : ::react::uwp::IReactInstanceCreator
-{
-  InstanceReactInstanceCreator(Microsoft::WRL::ComPtr<ABI::react::uwp::IInstance>& outer)
-  {
-    outer.AsWeak(&m_wrInstance);
-  }
-
-  std::shared_ptr<::react::uwp::IReactInstance> getInstance()
-  {
-    Microsoft::WRL::ComPtr<ABI::react::uwp::IInstance> spInstance;
-    m_wrInstance.As(&spInstance);
-    if (!spInstance)
-      return nullptr;
-    return reinterpret_cast<Instance*>(spInstance.Get())->getInstance();
-  }
-
-  void markAsNeedsReload()
-  {
-    Microsoft::WRL::ComPtr<ABI::react::uwp::IInstance> spInstance;
-    m_wrInstance.As(&spInstance);
-    reinterpret_cast<Instance*>(spInstance.Get())->markAsNeedsReload();
-  }
-
-private:
-  Microsoft::WRL::WeakRef m_wrInstance;
-  Microsoft::WRL::ComPtr<ABI::react::uwp::IInstance> m_instance;
-};
-
-Instance::Instance(HSTRING bundleFileName)
-{
-  m_jsBundleName = HSTRINGToString(bundleFileName);
-  Microsoft::WRL::ComPtr<IInstance> spThis(this);
-  m_instanceCreator = std::make_shared<InstanceReactInstanceCreator>(spThis);
-}
-
-Instance::Instance(const ::react::uwp::ReactInstanceCreator& instanceCreator)
-{
-  m_outerInstanceCreator = instanceCreator;
-  Microsoft::WRL::ComPtr<IInstance>spThis(this);
-  m_instanceCreator = std::make_shared<InstanceReactInstanceCreator>(spThis);
-}
-
-std::shared_ptr<::react::uwp::IReactInstance> Instance::getInstance()
-{
-  if (m_outerInstanceCreator)
-    return m_outerInstanceCreator->getInstance();
-
-  if (!m_instance)
-  {
-    m_instance = ::react::uwp::CreateReactInstance(m_spModuleProvider /*moduleLoader*/);
-    ::react::uwp::ReactInstanceSettings innerSettings;
-    innerSettings.UseLiveReload = m_settings.UseLiveReload;
-    innerSettings.UseWebDebugger = m_settings.UseWebDebugger;
-    m_instance->Start(m_instance, innerSettings);
-    m_instance->loadBundle(std::string(m_jsBundleName));
-  }
-  return m_instance;
-}
-
-void Instance::markAsNeedsReload()
-{
-  if (m_outerInstanceCreator)
-    return m_outerInstanceCreator->markAsNeedsReload();
-
-  m_instance->SetAsNeedsReload();
-  m_instance = nullptr;
-}
-
-HRESULT Instance::Start(ABI::react::uwp::InstanceSettings settings)
-{
-  if (m_outerInstanceCreator || m_instance)
-    return E_FAIL;
-
-  m_settings.UseLiveReload = settings.UseLiveReload;
-  m_settings.UseWebDebugger = settings.UseWebDebugger;
-  getInstance();
-
-  return S_OK;
-}
-
-HRESULT Instance::RegisterModule(ABI::react::uwp::IModule* pModule)
-{
-  if (m_outerInstanceCreator || m_instance)
-    return E_FAIL;
-
-  if (m_spModuleProvider == nullptr)
-    m_spModuleProvider = std::make_shared<ABIModuleLoader>();
-
-  Microsoft::WRL::ComPtr<ABI::react::uwp::IModule> spModule(pModule);
-  m_spModuleProvider->RegisterModule(spModule);
-  return S_OK;
-}
-
-const ::react::uwp::ReactInstanceCreator& Instance::GetReactInstanceCreator()
-{
-  return m_instanceCreator;
-}
-
-HRESULT InstanceStatics::Create(_In_ HSTRING bundleFileName, _Outptr_ ABI::react::uwp::IInstance** ppInstance)
-{
-  auto instance = Microsoft::WRL::Make<Instance>(bundleFileName);
-  return instance.CopyTo(ppInstance);
-}
-
 
 ReactControl::ReactControl()
   : m_rendering(false)
@@ -138,8 +33,6 @@ ReactControl::ReactControl()
   spFactory->CreateInstance(static_cast<ABI::react::uwp::IReactControl*>(this), &_spInnerInspectable, &spControl);
   SetComposableBasePointers(_spInnerInspectable.Get(), spFactory.Get());
 }
-
-
 
 HRESULT ReactControl::get_Instance(ABI::react::uwp::IInstance** value)
 {
@@ -178,7 +71,6 @@ HRESULT ReactControl::put_InitialProps(HSTRING value)
   m_initialProps = HSTRINGToString(value);
   return S_OK;
 }
-
 
 HRESULT ReactControl::StartRender()
 {
