+++ RNWCPP/ReactUWP/ABI/idl/Instance.idl
@@ -1,0 +1,50 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import "inspectable.idl";
+
+import "Module.idl";
+
+#define VERSION 1.0
+
+namespace react{
+  namespace uwp {
+
+  runtimeclass ReactInstance;
+
+  [version(VERSION)]
+  struct InstanceSettings
+  {
+      boolean	UseWebDebugger;
+      boolean	UseLiveReload;
+  };
+
+  [version(VERSION)]
+  [uuid(4A7C4023-0A50-4599-97A6-095528319E71)]
+  interface IInstance : IInspectable
+  {
+    HRESULT Start([in] InstanceSettings settings);
+
+    // We should really use a pull model like RN does -- but providing the push for now for ease of migration
+    HRESULT RegisterModule([in]IModule* pModule);
+  }
+
+  [exclusiveto(Instance)]
+  [version(VERSION)]
+  [uuid(A0B3AE3E-3E34-436D-BE4F-603A2C7098DE)]
+  interface IInstanceStatics : IInspectable
+  {
+    HRESULT Create([in] HSTRING bundleFileName, [out, retval] react.uwp.Instance** instance);
+  }
+
+  [version(VERSION)]
+  [marshaling_behavior(agile)]
+  [threading(both)]
+  [static(IInstanceStatics, VERSION)]
+  runtimeclass Instance
+  {
+    [default] interface IInstance;
+  };
+
+  }
+}
