+++ RNWCPP/ReactUWP/ABI/idl/Module.idl
@@ -1,0 +1,40 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import "inspectable.idl";
+import "Windows.Foundation.idl";
+
+#define VERSION 1.0
+
+namespace react{
+  namespace uwp {
+
+  [version(VERSION), uuid(264B6078-B38B-4C86-A796-8D2B1BC5CBCA)]
+  delegate HRESULT MethodDelegate([in]HSTRING args);
+
+  [version(VERSION), uuid(9E1F6C38-0270-4F22-9B90-7DA312DB60B9)]
+  delegate HRESULT MethodCallback(Windows.Foundation.Collections.IVectorView<HSTRING>* pCallBack);
+
+  [version(VERSION), uuid(C9ED1358-8239-4E71-BBF1-776200E841B2)]
+  delegate HRESULT MethodWithCallbackDelegate([in]HSTRING args, [in]MethodCallback* callBack);
+
+  declare
+  {
+    interface Windows.Foundation.Collections.IMapView<HSTRING, MethodDelegate*>;
+    interface Windows.Foundation.Collections.IMapView<HSTRING, MethodWithCallbackDelegate*>;
+  }
+
+  [version(VERSION), uuid(765C0317-7210-4E72-873B-CF0384B2B606)]
+  interface IModule : IInspectable
+  {
+    [propget] HRESULT Name([out][retval] HSTRING* value);
+
+    [propget] HRESULT Constants([out][retval] Windows.Foundation.Collections.IMapView<HSTRING, HSTRING>** value);
+
+    [propget] HRESULT Methods([out][retval] Windows.Foundation.Collections.IMapView<HSTRING, MethodDelegate*>** value);
+
+    [propget] HRESULT MethodsWithCallback([out][retval] Windows.Foundation.Collections.IMapView<HSTRING, MethodWithCallbackDelegate*>** value);
+  }
+
+  }
+}
