+++ RNWCPP/ReactUWP/ABI/idl/ReactControl.idl
@@ -1,0 +1,46 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+import "inspectable.idl";
+import "Windows.UI.Xaml.Controls.idl";
+
+import "Instance.idl";
+
+#define VERSION 1.0
+
+namespace react{
+  namespace uwp {
+
+  runtimeclass ReactControl;
+
+  [version(VERSION), uuid(BA8CD9D5-78E7-451E-9FA6-FAEAE6B43C52)]
+  interface IReactControl : IInspectable
+  {
+    [propget] HRESULT JsComponentName([out][retval] HSTRING* value);
+    [propput] HRESULT JsComponentName([in] HSTRING value);
+    [propget] HRESULT InitialProps([out][retval] HSTRING* value);
+    [propput] HRESULT InitialProps([in] HSTRING value);
+
+    [propget] HRESULT Instance([out][retval] IInstance** value);
+    [propput] HRESULT Instance([in] IInstance* value);
+
+    HRESULT StartRender();
+    HRESULT StopRender();
+  }
+
+  [exclusiveto(ReactControl)]
+  [version(VERSION), uuid(B882C924-0A23-4B4A-9088-8FAECB2D5945)]
+  interface IReactControlFactory : IInspectable
+  {
+    HRESULT CreateInstance([in] IInspectable* outer,[out] IInspectable** inner,[out][retval]react.uwp.ReactControl** instance);
+  }
+
+  [composable(react.uwp.IReactControlFactory, public, VERSION)]
+  [version(VERSION), marshaling_behavior(agile), threading(both)]
+  runtimeclass ReactControl : Windows.UI.Xaml.Controls.Grid
+  {
+    [default] interface IReactControl;
+  };
+
+  }
+}
