+++ RNWCPP/ReactUWP/ABI/idl/sources
@@ -7,15 +7,22 @@ MDMERGE_COMPOSITION_DEPTH0 = 2
 MIDL_UNIQUE_DLLDATA_NAMES = 1	
 	
 
-SOURCES = reactcontrol.g.idl \
+SOURCES = \
+	module.idl \
+	instance.idl \
+	reactcontrol.idl \
 
 # Noop rules needed to tell nmake that the unmerged.winmd files are produced by
 # midl, which also produces the .h files. Otherwise it tries to parallelize and
 # runs mdmerge before midl has created the winmds
-$(OBJ_PATH)\$O\reactcontrol.g.unmerged.winmd : $(PROJINCDIR)\reactcontrol.g.h
+$(OBJ_PATH)\$O\module.unmerged.winmd : $(PROJINCDIR)\module.h
+$(OBJ_PATH)\$O\instance.unmerged.winmd : $(PROJINCDIR)\instance.h
+$(OBJ_PATH)\$O\reactcontrol.unmerged.winmd : $(PROJINCDIR)\reactcontrol.h
 
 MDMERGE_SOURCES0 = \
-	$(OBJ_PATH)\$O\reactcontrol.g.unmerged.winmd \
+	$(OBJ_PATH)\$O\module.unmerged.winmd \
+	$(OBJ_PATH)\$O\instance.unmerged.winmd \
+	$(OBJ_PATH)\$O\reactcontrol.unmerged.winmd \
 
 MDMERGE_RESULTS0 = \
 	React.UWP.winmd \
@@ -28,4 +35,6 @@ $(TARGETDIR)\React.UWP.winmd : $(OBJ_PATH)\$O\mdmerge\pass0\output\react.uwp.win
 	$(OMAKECOPY) $(OBJ_PATH)\$O\mdmerge\pass0\output\react.uwp.winmd $(TARGETDIR)\react.uwp.winmd
 
 COPYFILES_PASS0 = $(COPYFILES_PASS0); \
-	$(PROJINCDIR)\reactcontrol.g.h,$(TARGETINCDIR); \
+	$(PROJINCDIR)\module.h,$(TARGETINCDIR); \
+	$(PROJINCDIR)\instance.h,$(TARGETINCDIR); \
+	$(PROJINCDIR)\reactcontrol.h,$(TARGETINCDIR); \
