+++ RNWCPP/ReactUWP/ABI/lib/sources
@@ -5,4 +5,6 @@ SOURCES = $(SOURCES_LIB) \
 
 HEADERS = \
 	$(HEADERS) \
-	$(PROJINCDIR)\reactcontrol.g.h \
+	$(PROJINCDIR)\module.h \
+	$(PROJINCDIR)\instance.h \
+	$(PROJINCDIR)\reactcontrol.h \
