+++ RNWCPP/ReactUWP/ABI/sources.inc
@@ -5,5 +5,6 @@ LIBLETNAME = ReactUWP_ABI
 SOURCES_LIB = \
 	..\ABIModuleLoader.cpp \
 	..\HStringHelper.cpp \
+	..\Instance_rt.cpp \
 	..\ReactControl_rt.cpp \
