+++ RNWCPP/ReactUWP/ReactUWP.vcxproj
@@ -89,6 +89,9 @@
       <GenerateWindowsMetadata>true</GenerateWindowsMetadata>
       <GenerateDebugInformation>DebugFull</GenerateDebugInformation>
     </Link>
+    <Midl>
+      <AdditionalIncludeDirectories>$(ProjectDir)ABI\idl</AdditionalIncludeDirectories>
+    </Midl>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(OSS_RN)'=='true'">
     <ClCompile>
@@ -151,6 +154,7 @@
     <ClInclude Include="ABI\ABIModuleLoader.h" />
     <ClInclude Include="ABI\comTemplateLibrary.h" />
     <ClInclude Include="ABI\HStringHelper.h" />
+    <ClInclude Include="ABI\Instance_rt.h" />
     <ClInclude Include="ABI\ReactControl_rt.h" />
     <ClInclude Include="Base\UwpReactInstance.h" />
     <ClInclude Include="Executors\WebSocketJSExecutorUwp.h" />
@@ -214,6 +218,7 @@
     <ClCompile Include="..\Shared\MemoryTracker.cpp" />
     <ClCompile Include="ABI\ABIModuleLoader.cpp" />
     <ClCompile Include="ABI\HStringHelper.cpp" />
+    <ClCompile Include="ABI\Instance_rt.cpp" />
     <ClCompile Include="ABI\ReactControl_rt.cpp" />
     <ClCompile Include="..\Shared\InstanceManager.cpp" />
     <ClCompile Include="..\Shared\OInstance.cpp" />
@@ -314,7 +319,9 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <MidlRT Include="ABI\idl\ReactControl.g.idl" />
+    <MidlRT Include="ABI\idl\Module.idl" />
+    <MidlRT Include="ABI\idl\Instance.idl" />
+    <MidlRT Include="ABI\idl\ReactControl.idl" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
